old_files/requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bce-python-sdk==0.9.63
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.1.0
charset-normalizer==3.4.5
chromadb==1.5.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
durationpy==0.10
effdet==0.4.1
emoji==2.15.0
exceptiongroup==1.3.1
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.0
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-vision==3.12.1
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.4.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.6.0
humanfriendly==10.0
idna==3.11
imagesize==2.0.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
langchain==1.2.12
langchain-core==1.2.18
langdetect==1.0.9
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
modelscope==1.34.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.3
numba==0.64.0
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
paddleocr==3.4.0
paddlex==3.4.2
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pi_heif==1.3.0
pikepdf==10.5.0
pillow==12.1.1
prettytable==3.17.0
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.2
pyparsing==3.3.2
pypdf==6.8.0
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
pytz==2026.1.post1
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.3.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
unstructured==0.18.32
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0